requirements.txt
Flask==1.1.2
numpy==1.19.2
opencv_python==4.4.0.46
Pillow==9.0.0
scikit_learn==1.0.2
tensorflow==2.7.0